move get_copilot_versions and get_aws_versions to config domain and tests
